Создает компонент ProfileLink
