+++ src/components/header-nav/header-nav.tsx
@@ -1,19 +1,13 @@
 import { Link } from 'react-router-dom';
 import { AppRoute, AuthorizationStatus } from '../../constants';
 import { useAppDispatch, useAppSelector } from '../../hooks';
-import {
-  selectAuthStatus,
-  selectUser,
-} from '../../store/user-process/selectors';
+import { selectAuthStatus } from '../../store/user-process/selectors';
 import { logoutAction } from '../../store/api-actions';
-import { selectFavorites } from '../../store/favorites-data/selectors';
+import ProfileLink from '../profile-link/profile-link';
 
-function HeaderNav() {
+function HeaderNav(): JSX.Element {
   const dispatch = useAppDispatch();
   const authorizationStatus = useAppSelector(selectAuthStatus);
-  const user = useAppSelector(selectUser);
-  const favorites = useAppSelector(selectFavorites);
-
   const isAuth = authorizationStatus === AuthorizationStatus.Auth;
 
   const signInLink = (
@@ -26,22 +20,11 @@ function HeaderNav() {
     </Link>
   );
 
-  const profileLink = (
-    <Link
-      className="header__nav-link header__nav-link--profile"
-      to={AppRoute.Favorites}
-    >
-      <div className="header__avatar-wrapper user__avatar-wrapper" />
-      <span className="header__user-name user__name">{user?.email}</span>
-      <span className="header__favorite-count">{favorites.length}</span>
-    </Link>
-  );
-
   return (
     <nav className="header__nav">
       <ul className="header__nav-list">
         <li className="header__nav-item user">
-          {!isAuth ? signInLink : profileLink}
+          {!isAuth ? signInLink : <ProfileLink />}
         </li>
         {isAuth && (
           <li className="header__nav-item">
